refactor(toolkit-lib): move payloads out of api
The message payload interfaces never were an api. They just ended up there by chance. Move them to their own top-level module. This has the side-effect that it will make future diffs slightly nicer.
