+++ packages/@aws-cdk/tmp-toolkit-helpers/src/api/io/private/messages.ts
@@ -1,20 +1,20 @@
 import type * as cxapi from '@aws-cdk/cx-api';
 import * as make from './message-maker';
 import type { SpanDefinition } from './span';
-import type { DiffResult } from '../payloads';
-import type { BootstrapEnvironmentProgress } from '../payloads/bootstrap-environment-progress';
-import type { MissingContext, UpdatedContext } from '../payloads/context';
-import type { BuildAsset, DeployConfirmationRequest, PublishAsset, StackDeployProgress, SuccessfulDeployStackResult } from '../payloads/deploy';
-import type { StackDestroy, StackDestroyProgress } from '../payloads/destroy';
-import type { HotswapDeploymentDetails, HotswapDeploymentAttempt, HotswappableChange, HotswapResult } from '../payloads/hotswap';
-import type { StackDetailsPayload } from '../payloads/list';
-import type { CloudWatchLogEvent, CloudWatchLogMonitorControlEvent } from '../payloads/logs-monitor';
-import type { StackRollbackProgress } from '../payloads/rollback';
-import type { SdkTrace } from '../payloads/sdk-trace';
-import type { StackActivity, StackMonitoringControlEvent } from '../payloads/stack-activity';
-import type { StackSelectionDetails } from '../payloads/synth';
-import type { AssemblyData, ConfirmationRequest, ContextProviderMessageSource, Duration, ErrorPayload, StackAndAssemblyData } from '../payloads/types';
-import type { FileWatchEvent, WatchSettings } from '../payloads/watch';
+import type { DiffResult } from '../../../payloads';
+import type { BootstrapEnvironmentProgress } from '../../../payloads/bootstrap-environment-progress';
+import type { MissingContext, UpdatedContext } from '../../../payloads/context';
+import type { BuildAsset, DeployConfirmationRequest, PublishAsset, StackDeployProgress, SuccessfulDeployStackResult } from '../../../payloads/deploy';
+import type { StackDestroy, StackDestroyProgress } from '../../../payloads/destroy';
+import type { HotswapDeploymentDetails, HotswapDeploymentAttempt, HotswappableChange, HotswapResult } from '../../../payloads/hotswap';
+import type { StackDetailsPayload } from '../../../payloads/list';
+import type { CloudWatchLogEvent, CloudWatchLogMonitorControlEvent } from '../../../payloads/logs-monitor';
+import type { StackRollbackProgress } from '../../../payloads/rollback';
+import type { SdkTrace } from '../../../payloads/sdk-trace';
+import type { StackActivity, StackMonitoringControlEvent } from '../../../payloads/stack-activity';
+import type { StackSelectionDetails } from '../../../payloads/synth';
+import type { AssemblyData, ConfirmationRequest, ContextProviderMessageSource, Duration, ErrorPayload, StackAndAssemblyData } from '../../../payloads/types';
+import type { FileWatchEvent, WatchSettings } from '../../../payloads/watch';
 
 /**
  * We have a rough system by which we assign message codes:
