+++ packages/@aws-cdk/tmp-toolkit-helpers/src/index.ts
@@ -1,2 +1,3 @@
 export * from './api';
+export * from './payloads';
 export * from './util';
