+++ packages/@aws-cdk/tmp-toolkit-helpers/src/payloads/bootstrap-environment-progress.ts
