+++ packages/@aws-cdk/tmp-toolkit-helpers/src/payloads/destroy.ts
