+++ packages/@aws-cdk/tmp-toolkit-helpers/src/payloads/hotswap.ts
@@ -1,7 +1,7 @@
 import type { PropertyDifference, Resource } from '@aws-cdk/cloudformation-diff';
 import type * as cxapi from '@aws-cdk/cx-api';
 import type { Duration } from './types';
-import type { ResourceMetadata } from '../../resource-metadata/resource-metadata';
+import type { ResourceMetadata } from '../api/resource-metadata/resource-metadata';
 
 /**
  * A resource affected by a change
