+++ packages/@aws-cdk/tmp-toolkit-helpers/src/payloads/progress.ts
