+++ packages/@aws-cdk/tmp-toolkit-helpers/src/payloads/sdk-trace.ts
