+++ packages/@aws-cdk/tmp-toolkit-helpers/src/payloads/stack-details.ts
