+++ packages/@aws-cdk/tmp-toolkit-helpers/src/private/activity-printer/current.ts
@@ -3,7 +3,7 @@ import * as chalk from 'chalk';
 import type { ActivityPrinterProps } from './base';
 import { ActivityPrinterBase } from './base';
 import { RewritableBlock } from './display';
-import type { StackActivity } from '../../api/io/payloads';
+import type { StackActivity } from '../../payloads';
 import { padLeft, padRight, stackEventHasErrorMessage } from '../../util';
 
 /**
