+++ packages/@aws-cdk/tmp-toolkit-helpers/src/private/activity-printer/history.ts
@@ -2,7 +2,7 @@ import * as util from 'util';
 import * as chalk from 'chalk';
 import type { ActivityPrinterProps } from './base';
 import { ActivityPrinterBase } from './base';
-import type { StackActivity } from '../../api/io/payloads';
+import type { StackActivity } from '../../payloads';
 import { padRight } from '../../util';
 
 /**
