+++ packages/@aws-cdk/toolkit-lib/lib/api/shared-public.ts
@@ -1,3 +1,5 @@
 /* eslint-disable import/no-restricted-paths */
 
 export * from '../../../tmp-toolkit-helpers/src/api';
+
+export * from '../../../tmp-toolkit-helpers/src/payloads';
