+++ packages/aws-cdk/lib/api/deployments/hotswap-deployments.ts
@@ -5,10 +5,10 @@ import type { WaiterResult } from '@smithy/util-waiter';
 import * as chalk from 'chalk';
 import type { SuccessfulDeployStackResult } from './deployment-result';
 import { ToolkitError } from '../../../../@aws-cdk/tmp-toolkit-helpers/src/api';
-import type { AffectedResource, HotswapResult, ResourceSubject, ResourceChange, NonHotswappableChange } from '../../../../@aws-cdk/tmp-toolkit-helpers/src/api/io/payloads';
-import { NonHotswappableReason } from '../../../../@aws-cdk/tmp-toolkit-helpers/src/api/io/payloads';
 import type { IMessageSpan, IoHelper } from '../../../../@aws-cdk/tmp-toolkit-helpers/src/api/io/private';
 import { IO, SPAN } from '../../../../@aws-cdk/tmp-toolkit-helpers/src/api/io/private';
+import type { AffectedResource, HotswapResult, ResourceSubject, ResourceChange, NonHotswappableChange } from '../../../../@aws-cdk/tmp-toolkit-helpers/src/payloads';
+import { NonHotswappableReason } from '../../../../@aws-cdk/tmp-toolkit-helpers/src/payloads';
 import { formatErrorMessage } from '../../util';
 import type { SDK, SdkProvider } from '../aws-auth';
 import type { CloudFormationStack, NestedStackTemplates } from '../cloudformation';
