+++ packages/aws-cdk/lib/api/hotswap/code-build-projects.ts
@@ -3,7 +3,7 @@ import {
   type HotswapChange,
   classifyChanges,
 } from './common';
-import type { ResourceChange } from '../../../../@aws-cdk/tmp-toolkit-helpers/src/api/io/payloads/hotswap';
+import type { ResourceChange } from '../../../../@aws-cdk/tmp-toolkit-helpers/src/payloads/hotswap';
 import { lowerCaseFirstCharacter, transformObjectKeys } from '../../util';
 import type { SDK } from '../aws-auth';
 import type { EvaluateCloudFormationTemplate } from '../cloudformation';
