+++ packages/aws-cdk/lib/api/hotswap/stepfunctions-state-machines.ts
@@ -1,5 +1,5 @@
 import { type HotswapChange, classifyChanges } from './common';
-import type { ResourceChange } from '../../../../@aws-cdk/tmp-toolkit-helpers/src/api/io/payloads/hotswap';
+import type { ResourceChange } from '../../../../@aws-cdk/tmp-toolkit-helpers/src/payloads/hotswap';
 import type { SDK } from '../aws-auth';
 import type { EvaluateCloudFormationTemplate } from '../cloudformation';
