+++ packages/aws-cdk/lib/api/logs/logs-monitor.ts
@@ -2,9 +2,9 @@ import * as util from 'util';
 import type * as cxapi from '@aws-cdk/cx-api';
 import * as chalk from 'chalk';
 import * as uuid from 'uuid';
-import type { CloudWatchLogEvent } from '../../../../@aws-cdk/tmp-toolkit-helpers/src/api/io';
 import type { IoHelper } from '../../../../@aws-cdk/tmp-toolkit-helpers/src/api/io/private';
 import { IO } from '../../../../@aws-cdk/tmp-toolkit-helpers/src/api/io/private';
+import type { CloudWatchLogEvent } from '../../../../@aws-cdk/tmp-toolkit-helpers/src/payloads';
 import { flatten } from '../../util';
 import type { SDK } from '../aws-auth';
